Add a CODEOWNERS file (#10)
# Motivations
This is completely optional, just a suggestion if you weren't aware of
this feature of GitHub. GitHub has documentation on this at:
https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners

I had an issue as I started to make projects that got some usage that by
default PRs and Issues weren't floating up to me. A `.github/CODEOWNERS`
file automatically assigns you when there are issues or PRs open, and
then these show up at https://github.com/pulls/assigned and
https://github.com/issues/assigned.

Feel free to close this if this isn't something you want!

# Modifications
- Adds a CODEOWNERS file