fix(configmap): daily-reboot: backup isn't a palworld rcon command (#15)
# Motivations
This fixes an issue where the `backup` command is used via rcon, which
doesn't exist.

# Modifications
- This PR fixes the above issue and adds a `set -euo pipefail` line to
the script to help catch things like this. This has the following
effects:
- The `set -e` option instructs bash to immediately exit if any command
returns a non-zero exit status, which will mark the pod as failed so k8s
can handle it appropriately. This would have caught the `backup` command
being sent to rcon
- `set -u` affects variables. When set, a reference to any variable you
haven't previously defined will error. This is useful for catching
variable typos, this caught an issue I describe below
- `set -o pipefail` turns on the `pipefail` option which prevents errors
in a pipeline from being masked. If any command in a pipeline fails,
that return code will be used as the return code of the whole pipeline.
For example `exit 1 | sort` would continue execution without `set -o
pipefail` when it should be treated as an error.
- `set -u` found that interpolation wasn't quite working on the
countdown: `i_seconds: unbound variable`. This caused the countdown to
not broadcast to players correctly. This PR fixes this