feat: allow broader storage configuration (#25)
Fixes #12 

# Motivations
Discussed in #12

# Modifications

This PR changes the structure of the `server.storage` block and is
therefore a breaking change.

This takes the existing storage config (below for reference):
```
server:
  # Storage definitions related to the palworld-server
  #
  storage:
    external: false
    externalName: ""

    # Keeps helm from deleting the PVC, by default helm does not delete pvcs
    #
    preventDelete: false

    size: 12Gi
    storageClassName: ""
```

and modifies it to the following, basically moving it under `main`.

```
server:
  # Storage definitions related to the palworld-server
  #
  storage:
    main:
      external: false
      externalName: ""

      # Keeps helm from deleting the PVC, by default helm does not delete pvcs
      #
      preventDelete: false

      size: 12Gi
      storageClassName: ""
```

---

Then we have the `backups` key, which supports any Kubernetes [Volume
spec](https://kubernetes.io/docs/reference/kubernetes-api/config-and-storage-resources/volume/#Volume)
as it's body. The `persistentVolumeClaim` is a special case which
matches the schema of `server.storage.main` rather than the Kubernetes
Volume spec

```
server:
  storage:
    backups: {}
      # persistentVolumeClaim:
      #   external: false
      #   externalName: ""
      #   preventDelete: false
      #   size: 12Gi
      #   storageClassName: ""

      # nfs:
      #   server: your.server.ip
      #   path: /path/to/remote/dir
```

---

Finally, we have the `extra` array. This takes a list of objects where
`name` and `mountPath` are required, along with a Kubernetes volume
spec. In `extra`, the `subPath` and `readOnly` of the volume mounts on
the deployment are configurable. If a `persistentVolumeClaim` is used,
`external` can be set to not provision a new PVC, otherwise `size`,
`storageClassName`, and `preventDelete` can be used to modify the
provisioned PVC. The `values.yaml` shows several commented examples of
this:
<https://github.com/Twinki14/palworld-server-chart/blob/5fae7415b4228c05a1ebc091e4e032342d65f9ea/charts/palworld/values.yaml#L45-L94>